Add Content to Schedule Page
Add the known schedule content to the schedule page based on what Cait
sent to Kurtis.

To adjust the centering, look at the '#event-table td' styles in
css/rp_styles.css

It would be a lot easier to center the content if we used <div>s or
<ul>s instead of the table.  Part of the problem right now is that we
cannot center the block of events with the day heading.  Content seems
weighted to the right because the event description tends to be longer
than the event time.